Start engine/src/visualization/ split: point cloud + mesh geometry helpers
Moves the round-point-texture cache, point material optimization, and
optimized-point-cloud creation into visualization/PointCloudRenderer.ts,
and the BufferGeometry-building logic (createGeometryFromSpatialData)
into visualization/MeshBuilder.ts, following the same host-interface +
thin-delegating-wrapper pattern used throughout main.ts's earlier
extraction pass. Also drops an inert debug block (computed bounding-box
stats that were never used) uncovered while moving the method.

main.ts: 4,351 -> 4,128 lines. This is a first step toward the
CLAUDE.md-proposed visualization/ structure (SceneManager, LightingManager
still to come) rather than the full split, since main.ts's remaining
scene lifecycle code (init/animate/dispose/context-loss) is far more
state-entangled and higher-risk to move without a dedicated pass.

Verified with tsc, full webpack build, and the full Playwright suite
(21/21 passing, including the PLY end-to-end load test that exercises
both extracted code paths directly).
